BOJ: 2단계(조건문) 문제 풀이 완료
