+++ wkdan_Java/src/Main/java/BOJ_02/p2753.java
@@ -1,4 +1,4 @@
-package BJ_02;
+package BOJ_02;
 
 import java.io.*;
