+++ wkdan_Java/src/Main/java/BOJ_02/p2884.java
@@ -1,24 +1,24 @@
-package BJ_02;
+package BOJ_02;
 
 import java.io.*;
 import java.util.StringTokenizer;
 
-public class p1330 {
+public class p2884 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         StringTokenizer st = new StringTokenizer(br.readLine());
+        int H = Integer.parseInt(st.nextToken());
+        int M = Integer.parseInt(st.nextToken());
 
-        int A = Integer.parseInt(st.nextToken());
-        int B = Integer.parseInt(st.nextToken());
+        int x = H * 60 + M;
+        int y = x - 45;
 
-        if (A < B) {
-            bw.write("<" + "\n");
-        } else if (A > B) {
-            bw.write(">" + "\n");
+        if (H == 0 && M <45) {
+            bw.write("23 "+ (M+15)+"\n");
         } else {
-            bw.write("==" + "\n");
+            bw.write((y/60) + " " + Math.abs(y%60)+"\n");
         }
         bw.flush();
         bw.close();
